requirements.txt
numpy==1.26.4
pandas==2.2.3
torch @ https://download.pytorch.org/whl/cu124/torch-2.5.1%2Bcu124-cp311-cp311-linux_x86_64.whl
transformers==4.51.1
tqdm==4.67.1
gym==0.25.2
faiss-cpu==1.11.0
matplotlib==3.7.5

